# conduit-api/openapi-spec.yml
openapi: 3.0.3
info:
  title: conduit
  description: Backend API for Conduit Interface
  version: '@project.version@'
servers:
  - url: /api/v1
paths:
  /users:
    post:
      description: Register a new user
      operationId: createUser
      tags:
        - user
      responses:
        201:
          $ref: '#/components/responses/UserResponse'
        400:
          $ref: '#/components/responses/BadRequest'
        403:
          $ref: '#/components/responses/Error'
        404:
          $ref: '#/components/responses/Error'
      requestBody:
        required: true
        description: Details of the new user to register
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUserRequest'
  /users/login:
    post:
      operationId: authenticateUser
      summary: Existing user login
      tags:
        - user
      requestBody:
        description: User authentication information
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthenticationRequest'
      responses:
        200:
          $ref: '#/components/responses/UserResponse'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Error'
...
components:
  responses:
    UserResponse:
      description: User information
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UserResponse'
...
    Error:
      description: An error occurred.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: The request has not been processed, request contained bad or unexpected parameter
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    UserResponse:
      description: User authentication info.
      type: object
      required:
        - user
      properties:
        user:
          $ref: '#/components/schemas/User'
    User:
      description: User authentication info.
      type: object
      required:
        - bio
        - email
        - image
        - token
        - username
      properties:
        email:
          type: string
        token:
          type: string
        username:
          type: string
        bio:
          type: string
        image:
          type: string
    AuthenticationRequest:
      description: Authentication request body.
      type: object
      required:
        - user
      properties:
        user:
          $ref: '#/components/schemas/AuthenticationRequestDetails'
    AuthenticationRequestDetails:
      description: Authentication request details
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    NewUserRequest:
      description: Request body for user registration.
      type: object
      required:
        - user
      properties:
        user:
          $ref: '#/components/schemas/NewUser'
    NewUser:
      description: Request body for user registration.
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
...
    Error:
      description: An error.
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
    ErrorResponse:
      type: object
      description: Response container for the error scenario
      required:
        - transactionId
        - error
      properties:
        transactionId:
          type: string
        error:
          $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      description: Details for the error scenario
      required:
        - code
        - reason
        - message
        - messageId
      properties:
        code:
          type: number
          description: HTTP status code for the error scenario
        reason:
          $ref: '#/components/schemas/ErrorReason'
        messages:
          type: array
          description: Description of the error scenario
          items:
            type: string
        messageId:
          $ref: '#/components/schemas/ErrorMessageIdentifier'
    ErrorReason:
      type: string
      description: The generalized reason for the specific error scenarios
      enum: [ API_BAD_PARAM, API_UNAUTHORISED, API_NOT_FOUND, API_SERVICE_NOT_AVAILABLE, API_ERROR, API_THROTTLED ]
    ErrorMessageIdentifier:
      type: string
      description: Error message identifier for specific error scenarios
      enum: [ API_ING_MSGEID_00001, API_ING_MSGEID_00002,
              API_ING_MSGEID_00003, API_ING_MSGEID_00004,
              API_ING_MSGEID_00005, API_ING_MSGEID_00006,
              API_ING_MSGEID_00007, API_ING_MSGEID_00008 ]
  securitySchemes:
    Token:
      type: apiKey
      in: header
      name: Authorization